backend/requirements.txt
# NCHU AI Counselor API - FastAPI Backend

# Core framework
fastapi>=0.115.0
# Production server
gunicorn>=21.2.0
uvicorn[standard]>=0.32.0

# Configuration
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Cache
redis>=5.2.0

# Vector store
qdrant-client>=1.12.0

# HTTP client
httpx>=0.28.0

# HTML parsing
beautifulsoup4>=4.12.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0

# Linting and formatting
ruff>=0.8.0

# Structured logging
structlog>=24.4.0

# Authentication
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
